Move PFM to middlewares
